requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
pyyaml==6.0.1
great-expectations==0.18.10
lifetimes==0.11.3
lightgbm==4.3.0
shap==0.45.1
matplotlib==3.8.4
mlflow==2.13.0
flask==3.0.3
streamlit==1.36.0
plotly==5.22.0
joblib==1.4.2